chore: separate child components out for readability
